+++ rules/standard/src/main/kotlin/dev/s7a/sqldelight/check/rules/standard/StandardRuleSetProvider.kt
@@ -15,13 +15,17 @@ import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceAfterOpeningParenthe
 import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceBeforeClosingParenthesisRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceBeforeCommaRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceBeforeDotRule
+import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceBeforeFunctionParenthesisRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoSpaceBeforeSemicolonRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoTabIndentationRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoTrailingBlankLinesRule
 import dev.s7a.sqldelight.check.rules.standard.rules.NoTrailingWhitespaceRule
+import dev.s7a.sqldelight.check.rules.standard.rules.SpaceAfterBlockCommentStartRule
 import dev.s7a.sqldelight.check.rules.standard.rules.SpaceAfterCommaRule
 import dev.s7a.sqldelight.check.rules.standard.rules.SpaceAfterLineCommentMarkerRule
+import dev.s7a.sqldelight.check.rules.standard.rules.SpaceAroundBinaryOperatorsRule
 import dev.s7a.sqldelight.check.rules.standard.rules.SpaceAroundComparisonOperatorsRule
+import dev.s7a.sqldelight.check.rules.standard.rules.SpaceBeforeBlockCommentEndRule
 import dev.s7a.sqldelight.check.rule.api.RuleProvider
 import dev.s7a.sqldelight.check.rule.api.RuleSetProvider
 
@@ -51,12 +55,16 @@ public class StandardRuleSetProvider : RuleSetProvider {
             RuleProvider(::NoSpaceBeforeClosingParenthesisRule),
             RuleProvider(::NoSpaceBeforeCommaRule),
             RuleProvider(::NoSpaceBeforeDotRule),
+            RuleProvider(::NoSpaceBeforeFunctionParenthesisRule),
             RuleProvider(::NoSpaceBeforeSemicolonRule),
             RuleProvider(::NoTabIndentationRule),
             RuleProvider(::NoTrailingBlankLinesRule),
             RuleProvider(::NoTrailingWhitespaceRule),
+            RuleProvider(::SpaceAfterBlockCommentStartRule),
             RuleProvider(::SpaceAfterCommaRule),
             RuleProvider(::SpaceAfterLineCommentMarkerRule),
+            RuleProvider(::SpaceAroundBinaryOperatorsRule),
             RuleProvider(::SpaceAroundComparisonOperatorsRule),
+            RuleProvider(::SpaceBeforeBlockCommentEndRule),
         )
 }